Changing the default upper and lower bounds
We were using +/- max double but this triggers a runtime assert in the visual c++ compiler with uniform distributions. We don't really need such big default bounds, so it has been changed to +/- max float, but stored in a double.

+++ src/Bounds.cpp
@@ -26,8 +26,10 @@
 
 namespace Unfit
 {
-static const double max_positive_num {std::numeric_limits<double>::max()};
-static const double max_negative_num {-std::numeric_limits<double>::max()};
+static const double max_positive_num
+    {static_cast<double>(std::numeric_limits<float>::max())};
+static const double max_negative_num
+    {-static_cast<double>(std::numeric_limits<float>::max())};
 
 Bounds::Bounds()
   : upper_bound_ {},
